WIP: FIX: hostNetwork support for E2 termination
This change implements Host Networking support for the E2 termination
container.  The intention is to provide a prototype implementation
supporting the container's networking requirements while allowing for
flexibility in the choice of that implementation; in particular, it
should ideally be possible to replace hostNetwork support with, for
example, DANM or Multus with no chart changes beyond values.yaml.

The particular network requirements for E2 termination which mandate
a special network configuration are:
o Multiple network interfaces (one for RMR and one for E2)
o SCTP on the E2 interface
o IPv6 on the E2 interface
o Communication with other pods via RMR the CNI network
o IPv4 on the CNI interface

The change also adds a generic initcontainer for RIC platform
components.

Change-Id: I92d4bc72a76fa1d0f795b051cee34e914ec791ab
Signed-off-by: dave kormann <dk3239@att.com>