Update patch set 1
Patch Set 1:

This change is COMPLETELY untested so far, and is intended at this point as a placeholder for a complete implementation.

For at least the hostnetwork case, a startup script or initcontainer will need to be added which either determines the address of the CNI bridge (if RMR's to bind to that) or adds a veth pair with one end plugged into the CNI bridge and the other exposed as the bind interface for the RMR implementation.  In the latter case, we'll want to revisit the proposed values.yaml additions.

Patch-set: 1
Reviewer: Gerrit User 1000047 <1000047@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=0