Update patch set 7
Patch Set 7:

The docker image is available on my VM, it's not yet in any registry.

The springApplicationJson is in-line configuration, it's not an external JSON file.  Can you show how other components are using configmaps so I do something compatible?  How do I specify a configmap in a helm chart?

Patch-set: 7
Label: Code-Review=0