Update patch set 12
Patch Set 12:

(1 comment)

> Uploaded patch set 12.


This will create runtime problem when the internal expose port is set to the same K8S service port. Dashboard is listening at 8080 by design. However, that port will be mapped to the service port at runtime. We may need to move the service port around later. So we should use a different port definition. If you want to also may it configurable, we need to add a .Values.container.port in the value.yaml file.

Patch-set: 12
Label: Code-Review=0