Modify ric_install script to install from local chart files. Reorganize folder structure.
The ric_install script is no longer setting up a local helm repo. Charts
are directly deployed form local files. This also allows us to delete
and redeploy individual components for debuging purpose. There is a plan
to move the helm repo creation and uploading helm chart script to the
aux directory as part of the PIZ scripts.

The project structure is also reorganized so that we have all the
platform related components in the platform folder, all xapp related
files go to xapp folder, and files that create the development
enviroment goes to dev folder.

Change-Id: I1e1a31e4538b4169afef3580d01907d5aa72a387
Signed-off-by: Zhe Huang <zhehuang@research.att.com>