Add namespace name to release name
Credentials are per namespace therefore we need to deploy this chart
for each namespace where the credential is needed.  The helm release
name used for deploying the chart to each namespace should reflect the
namespace; otherwise we will have "release already exists" error when
deploy the credential into the second namespace in the same cluster.

Change-Id: I8fb013900113f3f16ecff40b8c32edfd32679a06
Signed-off-by: wrider <lji@research.att.com>