Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> (3 comments)

The 1.0.0 is the chart version, which we are keeping consistent through all the charts.
I put the other components back into the components string.
The override file is defining a single registry.  After discussions with Lusheng, we pushed the image to azure.  Obviously only a stopgap measure.  As Chris mentioned, that should not be in the file, so I removed it in the latest patch.

Patch-set: 5