Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 4:
> > 
> > (3 comments)
> 
> The 1.0.0 is the chart version, which we are keeping consistent through all the charts.
> I put the other components back into the components string.
> The override file is defining a single registry.  After discussions with Lusheng, we pushed the image to azure.  Obviously only a stopgap measure.  As Chris mentioned, that should not be in the file, so I removed it in the latest patch.

Can you please elaborate on the Azure image push?

The reason this is important is because the version of this container is going to (and already has) changed; and the CICD jobs push to LF. So I already have a version bump to make here. 

I had a working secret for pulling from LF, which was removed here, so I assumed there is already a secret defined for the LF in common.

Patch-set: 5