Update patch set 2
Patch Set 2:

I suggest that the prefix should be the well-known ingress path.  It may be little silly to repeat it for each endpoint, but that does little harm.  And the suffix should have everything else.  For example

    a1med:
      url:
        prefix: 
        suffix: /a1mediator

I don't fully grasp the need for an "override" option.  The deployment must determine the path; the helm chart should never, ever override a deployment!

So if you go with the suggestion for prefix/suffix, and drop the override thing, then I see no need to do any string catenation in Mustache templates and this can be very simple.

But I'm awaiting your input to resolve the conflict.

Patch-set: 2
Reviewer: Gerrit User 1000014 <1000014@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=0