Update patch set 3
Patch Set 3:

So the values.yaml file is defining the default values to be used when a recipe is not provided. It is a good practice to provide as many default values as possible to minimize the chance of errors caused by bad recipe. Is it possible to provide default values to the URLs?

Patch-set: 3
Label: Code-Review=0