Update patch set 12
Patch Set 12: Code-Review+1

took the WIP tag off this because I think it's close to ready.  I'd like other folks to look at it.

it could use integration with the common configuration process, and we still need to configure appmgr to use it, but it's most of the way there.

Patch-set: 12
Label: Code-Review=+1