Update patch set 14
Patch Set 14:

> Dave please +1 when you think we should be testing this.

you can probably start in on looking at it now.  however, it's probably not quite functional enough to merge so i'll avoid +1'ing just yet.

to use it, you'll want to look at the changes to RECIPE_EXAMPLE/RIC_INFRA_RECIPE_EXAMPLE, specifically the tillers: subtree; each subtree of that specifies the configuration of a tiller instance in the namespace with the name of the subtree (so, for example, "tillers.ricxapp.secret.create.true" means create a secret in the ricxapp namespace).  the defaults should be ok for most purposes.  to install the tillers, do a helm install of 'ricplt-tiller' in ric-infra/25-tiller/helm.  if that works, you should be able to do an install of the appmgr.

there's one known problem right now: the appmgr chart doesn't have a configuration for the helm repo's CA certificate, so it can't talk.  but it _does_ seem to talk successfully to the new tiller.

Patch-set: 14
Label: Code-Review=0