Update patch set 1
Patch Set 1:

@Felix, this commit is to test the AAF and Portal helm charts as is to deploy in ric-aux. We did test them in a k8 env and worked. Please see, if you can help us test this commit and provide +2. Thanks.

For now, we want to keep them as is and merge the charts as a baseline and then we want to clean up further to remove any unnecessary stuff. So that we know what is the working state and can figure out if something breaks due to any clean up.

Patch-set: 1