Update patch set 2
Patch Set 2:

When the curl command works, the whole job works beautifully, and this time actually caught a problem in one of the charts.  But the curl time out problem is still there.  
Time to build a "dep verify" container with helm client already installed, then verify job just runs lint in it?

Patch-set: 2