Update patch set 3
Patch Set 3:

The rtmgr charts need further integration. Currently the deployment.yaml does not have a http port defined. Also the http port value in the common template is not 8888. Hence the nbiurl doesn't work 100%. 

However, I will merge this now since it does not conflict with the future changes. We will be completing this in the next patch.

Patch-set: 3