Update patch set 2
Patch Set 2:

My understanding is that the Kibana site gets configured manually, and when that's all done, you can have it generate a configuration name.  I think the challenge will be to export and/or freeze the config so it can be recreated upon deployment with exactly the same ID so the same URL can be used.

Patch-set: 2