Update patch set 1
Patch Set 1:

The idea is that now helm verify job only does verification of charts in the it/dep repo. The integration team also needs to make sure that the submodule recorded SHA-1 points to a correct version of ric-plt/ric-dep to form a deployable job.

Therefore, the packaging job should first verify this is the case before generating the package.

Patch-set: 1