NONRTRIC - PMS Persistent storage of policies and type definitions
Changed deployment to statefulset in both PMC and ECS.

"Even Deployments with one replica using ReadWriteOnce volume are not recommended. This is because the default Deployment strategy creates a second Pod before bringing down the first Pod
 on a recreate. The Deployment may fail in deadlock as the second Pod can't start because the ReadWriteOnce volume is already in use, and the first Pod won't be removed because the seco
nd Pod has not yet started. Instead, use a StatefulSet with ReadWriteOnce volumes."

Change-Id: If3afb358e9a111bf2be0d5ceb9c5a70c314a9222
Signed-off-by: PatrikBuhr <patrik.buhr@est.tech>
Issue-ID: NONRTRIC-478