NONRTRIC - PMS Fix dynamic configuration
The problem was that the config cannot be updated via REST if the file is controlled by a configmap.
The configmap for PMS application can only be used for initial configuration after this change.

The file is moved from
/opt/app/policy-agent/data/application_configuration.json -> /var/policy-management-service/application_configuration.json

/var/policy-management-service/ is mapped to a persistent volume claim.

Config map which now is only initial (installation) data is mapped to:
/etc/app/policy-management-service/initialdata

An init container copies the initial configuration file /etc/app/policy-management-service/initialdata/application_configuration.json to /var/policy-management-service/
unless it is already there.

Signed-off-by: PatrikBuhr <patrik.buhr@est.tech>
Issue-ID: NONRTRIC-689
Change-Id: I79be05bbc4d18ad1914efc3cbcdf932ede71718d