Default k8s ppnt whitelist added in onap-override.yaml
ONAP Policy K8s ppnt needs the helm repositories to be whitelisted. This needs to be done for all the new helm repositories to be used with ONAP Policy K8s ppnt.
This commits adds the default helm repositories for ONAP Policy K8s ppnt in the onap-override.yaml file.
Any valid helm repository can be added to this override before installation.

Issue-ID: INT-170
Change-Id: I25ace887c2b7010e7f9124cd7c1f29e0d6d7d53a
Signed-off-by: aravind.est <aravindhan.a@est.tech>