Fix - Focom adapter certificate issue
Focom adapter requires o2ims and focom kubeconfig files to run properly. It is also designed to not report any error when the kubeconfig controlplane is invalid or not reachable. But the fabric8 kubernetes client throws a runtime error when the certificate given is invalid.

It causes the application to restart indefinitely. This change is to add valid certificate, so that the application will not exit rather log the error to show that it is unable to reach the given kube apiserver

Issue-ID: SMO-206
Change-Id: I7ffec147b8c99ebd2438800bd637e1b77c5bc9c2
Signed-off-by: aravind.est <aravindhan.a@est.tech>